Fixing field size

+++ FFXIVClientStructs/FFXIV/Client/Game/Gauge/JobGauges.cs
@@ -149,7 +149,7 @@ public struct MonkGauge {
     [FieldOffset(0x0B)]
     public BeastChakraType BeastChakra3; // CoeurlChakra = 1, RaptorChakra = 2, OpoopoChakra = 3 (only one value)
 
-    [FieldOffset(0x0C)] public ushort BeastChakraStacks;
+    [FieldOffset(0x0C)] public byte BeastChakraStacks;
 
     [FieldOffset(0x0D)] public NadiFlags Nadi; // LunarNadi = 1, SolarNadi = 2 (If both then 1+2=3)
     [FieldOffset(0x0E)] public ushort BlitzTimeRemaining; // 20 seconds
